requirements.txt
transformers
accelerate
sentence-transformers
faiss-cpu
gradio